+++ src/dataProvider.ts
@@ -1,18 +1,18 @@
 import * as vscode from 'vscode';
-import { defaultTraceTemplate, defaultLayoutTemplate } from './previewTemplates';
-import { parseDocument, parseTextFromUri, SPEC_DATA_FILTER, DPPMCA_FILTER, DOCUMENT_SELECTOR, CSV_ROWS_FILTER } from './dataParser';
-import type { Node, ParserResult, ParserSuccess } from './dataParser';
+import { defaultDataTemplate, defaultLayoutTemplate } from './previewTemplates';
+import { parseDocument, parseTextFromUri, SPEC_DATA_FILTER, DPPMCA_FILTER, DOCUMENT_SELECTOR } from './dataParser';
+import type { Node, ParserResult, ParserSuccess, SupportedLanguage } from './dataParser';
 
 // @types/plotly.js contains DOM objects and thus
 // `tsc -p .` fails without `skipLibCheck`.
-import type { PlotData, Layout, Template } from 'plotly.js';
-// type PlotData = any;
-// type Layout = any;
+import type { Template } from 'plotly.js';
 import type { State, GraphMode, MessageToWebview, MessageFromWebview } from './previewTypes';
+import type { ColorThemeKindLabel } from './previewTemplates';
 
 type Preview = {
     readonly panel: vscode.WebviewPanel;
     uri: vscode.Uri;
+    language: SupportedLanguage;
     enableMultipleSelection: boolean;
     enableRightAxis: boolean;
     nodes?: Node[];
@@ -50,8 +50,7 @@ export class DataProvider implements vscode.FoldingRangeProvider, vscode.Documen
 
         this.colorThemeKind = vscode.window.activeColorTheme.kind;
 
-        const config = vscode.workspace.getConfiguration('spec-data.preview');
-        this.enablePreviewScroll = config.get<boolean>('scrollPreviewWithEditor', true);
+        this.enablePreviewScroll = vscode.workspace.getConfiguration('spec-data.preview').get<boolean>('scrollPreviewWithEditor', true);
 
         // Create a command handler function that shows a preview.
         const makeShowPreviewCallback = (lockPreview: boolean, showToSide: boolean) => {
@@ -252,6 +251,19 @@ export class DataProvider implements vscode.FoldingRangeProvider, vscode.Documen
                     preview.panel.webview.postMessage(messageOut);
                 }
             }
+            for (const preview of this.previews) {
+                const scope = { languageId: preview.language, uri: preview.uri };
+                if (
+                    event.affectsConfiguration('spec-data.preview.plot.colorScale', scope) ||
+                    event.affectsConfiguration('spec-data.preview.plot.template', scope)
+                ) {
+                    preview.panel.webview.postMessage({
+                        type: 'setTemplate',
+                        template: getPlotlyTemplate(this.colorThemeKind, scope),
+                        callback: 'relayout',
+                    } satisfies MessageToWebview);
+                }
+            }
         };
 
         const activeColorThemeChangeListener = (colorTheme: vscode.ColorTheme) => {
@@ -264,7 +276,7 @@ export class DataProvider implements vscode.FoldingRangeProvider, vscode.Documen
                     // However, it seems invisible webviews also handle the following messages.
                     preview.panel.webview.postMessage({
                         type: 'setTemplate',
-                        template: getPlotlyTemplate(colorTheme.kind, preview.uri),
+                        template: getPlotlyTemplate(colorTheme.kind, { uri: preview.uri, languageId: preview.language }),
                         callback: 'relayout',
                     } satisfies MessageToWebview);
                 }
@@ -338,15 +350,21 @@ export class DataProvider implements vscode.FoldingRangeProvider, vscode.Documen
             return;
         }
         const uri = vscode.Uri.parse(state.sourceUri);
-        const parsedData = await this.parseDataOfFileUri(uri);
-        if (!parsedData || parsedData.nodes === undefined) {
+        const parserResult = await this.parseDataOfFileUri(uri);
+        if (!parserResult || parserResult.nodes === undefined) {
             const message = vscode.l10n.t('Failed to parse file: "{0}"', vscode.workspace.asRelativePath(uri));
             vscode.window.showErrorMessage(message);
             return;
         }
 
-        const preview: Preview = { panel, uri, enableMultipleSelection: state.enableMultipleSelection, enableRightAxis: state.enableRightAxis, };
-        await this.postCreatePreview(preview, parsedData, state.lockPreview);
+        const preview: Preview = {
+            panel,
+            uri,
+            language: parserResult.language,
+            enableMultipleSelection: state.enableMultipleSelection,
+            enableRightAxis: state.enableRightAxis,
+        };
+        await this.postCreatePreview(preview, parserResult, state.lockPreview);
         return;
     }
 
@@ -390,8 +408,8 @@ export class DataProvider implements vscode.FoldingRangeProvider, vscode.Documen
             return this.livePreview;
         } else {
             // Else create a new webview panel.
-            const parsedData = await this.parseDataOfFileUri(uri);
-            if (!parsedData || parsedData.nodes === undefined) {
+            const parserResult = await this.parseDataOfFileUri(uri);
+            if (!parserResult || !parserResult.nodes) {
                 const message = vscode.l10n.t('Failed to parse file: "{0}"', vscode.workspace.asRelativePath(uri));
                 vscode.window.showErrorMessage(message);
                 return;
@@ -408,13 +426,17 @@ export class DataProvider implements vscode.FoldingRangeProvider, vscode.Documen
                     retainContextWhenHidden: config.get<boolean>('retainContextWhenHidden', false),
                 }
             );
-            const enableMultipleSelection = config.get<boolean>('plot.enableMultipleSelection', false);
-            const enableRightAxis = config.get<boolean>('plot.enableRightAxis', false);
-            const preview: Preview = { panel, uri, enableMultipleSelection, enableRightAxis };
+            const preview: Preview = {
+                panel,
+                uri,
+                language: parserResult.language,
+                enableMultipleSelection: config.get<boolean>('plot.enableMultipleSelection', false),
+                enableRightAxis: config.get<boolean>('plot.enableRightAxis', false)
+            };
             if (showToSide && config.get<boolean>('autoLockGroup', false)) {
                 vscode.commands.executeCommand('workbench.action.lockEditorGroup');
             }
-            return await this.postCreatePreview(preview, parsedData, lockPreview);
+            return await this.postCreatePreview(preview, parserResult, lockPreview);
         }
     }
 
@@ -428,11 +450,11 @@ export class DataProvider implements vscode.FoldingRangeProvider, vscode.Documen
     /**
      * Register event handlers and then make a content from source.
      * @param preview Preview object created during deserialization process or `showPreview`-type action command.
-     * @param parsedData ParsedData object
+     * @param parserResult ParserSuccess object
      * @param lockPreview flag to lock preview with source URI.
      * @returns Preview object if succeeded in parsing a file or `undefined`.
      */
-    private async postCreatePreview(preview: Preview, parsedData: ParserSuccess, lockPreview: boolean) {
+    private async postCreatePreview(preview: Preview, parserResult: ParserSuccess, lockPreview: boolean) {
         this.previews.push(preview);
         preview.panel.iconPath = new vscode.ThemeIcon('graph-line');
 
@@ -571,7 +593,8 @@ export class DataProvider implements vscode.FoldingRangeProvider, vscode.Documen
                 }
             } else if (messageIn.type === 'contentLoaded') {
                 // This will be called not only when the webview is created but also when it is revealed after it is hidden.
-                const config = vscode.workspace.getConfiguration('spec-data.preview');
+                const scope = { languageId: preview.language, uri: preview.uri };
+                const config = vscode.workspace.getConfiguration('spec-data.preview', scope);
 
                 preview.panel.webview.postMessage({
                     type: 'lockPreview',
@@ -585,13 +608,13 @@ export class DataProvider implements vscode.FoldingRangeProvider, vscode.Documen
 
                 preview.panel.webview.postMessage({
                     type: 'setTemplate',
-                    template: getPlotlyTemplate(vscode.window.activeColorTheme.kind, preview.uri),
+                    template: getPlotlyTemplate(vscode.window.activeColorTheme.kind, scope),
                     callback: 'newPlot',
                 } satisfies MessageToWebview);
 
                 preview.panel.webview.postMessage({
                     type: 'setScrollBehavior',
-                    value: config.get<boolean>('smoothScrolling', true) ? 'smooth' : 'auto'
+                    value: config.get<boolean>('smoothScrolling', false) ? 'smooth' : 'auto'
                 } satisfies MessageToWebview);
 
                 preview.panel.webview.postMessage({
@@ -601,7 +624,7 @@ export class DataProvider implements vscode.FoldingRangeProvider, vscode.Documen
             }
         }, null, this.subscriptions);
 
-        this.updatePreviewWithNodes(preview, parsedData.language, parsedData.nodes);
+        this.updatePreviewWithNodes(preview, parserResult.language, parserResult.nodes);
 
         return preview;
     }
@@ -619,15 +642,15 @@ export class DataProvider implements vscode.FoldingRangeProvider, vscode.Documen
             return preview;
         }
 
-        const parsedData = await this.parseDataOfFileUri(uri);
-        if (!parsedData || parsedData.nodes === undefined) {
+        const parserResult = await this.parseDataOfFileUri(uri);
+        if (!parserResult || !parserResult.nodes) {
             const message = vscode.l10n.t('Failed to parse file: "{0}"', vscode.workspace.asRelativePath(uri));
             vscode.window.showErrorMessage(message);
             return undefined;
         }
 
         preview.uri = uri;
-        this.updatePreviewWithNodes(preview, parsedData.language, parsedData.nodes);
+        this.updatePreviewWithNodes(preview, parserResult.language, parserResult.nodes);
         return preview;
     }
 
@@ -815,37 +838,44 @@ mode:
     </html>`;
 }
 
-type ColorThemeKind = 'light' | 'dark' | 'highContrast' | 'highContrastLight';
-
 function getPlotlyTemplate(kind: vscode.ColorThemeKind, scope?: vscode.ConfigurationScope): Template {
-    let traceTemplate: Partial<PlotData>[]; // Record<string, unknown>[];
-    let layoutTemplate: Partial<Layout>; // Record<string, unknown>;
-
     const config = vscode.workspace.getConfiguration('spec-data.preview.plot', scope);
-    const userTraceTemplate = config.get<{ [key in ColorThemeKind]?: Partial<PlotData>[] }>('traceTemplate');
-    const userLayoutTemplate = config.get<{ [key in ColorThemeKind]?: Partial<Layout> }>('layoutTemplate');
+    const colorscale = config.get<string>('colorScale', 'RdBu');
+    const userDataTemplate = config.get<{ [key in ColorThemeKindLabel]?: Template['data'] }>('template.data', {});
+    const userLayoutTemplate = config.get<{ [key in ColorThemeKindLabel]?: Template['layout'] }>('template.layout', {});
+    const additionalDataTemplate: Template['data'] = {
+        heatmap: [{ colorscale }], contour: [{ colorscale }],
+    };
 
+    let colorLabel: ColorThemeKindLabel;
     switch (kind) {
         case vscode.ColorThemeKind.Light:
-            traceTemplate = userTraceTemplate?.light ?? defaultTraceTemplate.light;
-            layoutTemplate = userLayoutTemplate?.light ?? defaultLayoutTemplate.light;
+            colorLabel = 'light';
             break;
         case vscode.ColorThemeKind.Dark:
-            traceTemplate = userTraceTemplate?.dark ?? defaultTraceTemplate.dark;
-            layoutTemplate = userLayoutTemplate?.dark ?? defaultLayoutTemplate.dark;
+            colorLabel = 'dark';
             break;
         case vscode.ColorThemeKind.HighContrast:
-            traceTemplate = userTraceTemplate?.highContrast ?? defaultTraceTemplate.highContrast;
-            layoutTemplate = userLayoutTemplate?.highContrast ?? defaultLayoutTemplate.highContrast;
+            colorLabel = 'highContrast';
             break;
         case vscode.ColorThemeKind.HighContrastLight:
-            traceTemplate = userTraceTemplate?.highContrastLight ?? defaultTraceTemplate.highContrastLight;
-            layoutTemplate = userLayoutTemplate?.highContrastLight ?? defaultLayoutTemplate.highContrastLight;
+            colorLabel = 'highContrastLight';
             break;
         default:
-            traceTemplate = [];
-            layoutTemplate = {};
+            // This case will not be hit. Return an empty template to satisfy the return type.
+            return {};
     }
 
-    return { data: { scatter: traceTemplate }, layout: layoutTemplate };
+    // Shallow merge the default template and user template.
+    const dataTemplate: NonNullable<Template['data']> =
+        colorLabel in userDataTemplate ?
+            { ...defaultDataTemplate[colorLabel], ...additionalDataTemplate, ...userDataTemplate[colorLabel] } :
+            { ...defaultDataTemplate[colorLabel], ...additionalDataTemplate };
+
+    const layoutTemplate: NonNullable<Template['layout']> =
+        colorLabel in userLayoutTemplate ?
+            { ...defaultLayoutTemplate[colorLabel], ...userLayoutTemplate[colorLabel] } :
+            defaultLayoutTemplate[colorLabel];
+
+    return { data: dataTemplate, layout: layoutTemplate };
 }